+++ Net_Core/DataAccess/Interfaces/ILogging.cs
@@ -1,7 +1,8 @@
 using GrowthWare.Framework.Models;
 using GrowthWare.DataAccess.Interfaces.Base;
-using System;
 using System.Data;
+using System.Collections.Generic;
+using System.Threading;
 
 namespace GrowthWare.DataAccess.Interfaces
 {
@@ -17,6 +18,13 @@ namespace GrowthWare.DataAccess.Interfaces
         /// <param name="logSeqId">int</param>
         /// <returns>MLoggingProfile</returns>
         MLoggingProfile GetLog(int logSeqId);
+
+        /// <summary>
+        /// Returns data from the Logging table
+        /// </summary>
+        /// <param name="cancellationToken"></param>
+        /// <returns>IAsyncEnumerable<IDataRecord></returns>
+        IAsyncEnumerable<IDataRecord> GetLogs(CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Saves this instance.